apps/ml/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlparse==0.4.4
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
celery==5.3.6
openai==1.10.0
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-mock==3.12.0
